[#880] Refactor publishing into convention plugin
